+++ ydb/core/formats/arrow/accessor/abstract/ya.make
@@ -8,6 +8,8 @@ PEERDIR(
     ydb/library/actors/core
     ydb/core/formats/arrow/accessor/common
     ydb/library/formats/arrow/protos
+    ydb/library/arrow_kernels
+    ydb/core/formats/arrow/program
 )
 
 SRCS(
